Implement InfiniteLoopThread, a generic thread that runs a function forever.
